refactor: rename elements.ts to places.ts for better clarity #276
- Rename src/lib/sync/elements.ts → src/lib/sync/places.ts
- Update import in src/routes/+layout.svelte to use new path
- Update comments referencing the old filename in documentation
- No functional changes - file now accurately reflects its purpose

The file syncs Place objects from v4 Places API, not Element objects,
so the new name better matches its actual functionality.
